1.02 after Q/A session code simplification new class for Scheduled readable way of finding tasks Table correction

+++ src/main/resources/liquibase/scripts/notifications.sql
@@ -3,7 +3,7 @@
 -- changeset denik:1
 CREATE TABLE notifications(
     id           SERIAL PRIMARY KEY,
-    chat_id      INTEGER,
+    chat_id      BIGINT,
     message      TEXT,
     date         TIMESTAMP
 )
